feat: Add isDemoUser middleware to check if user is a demo user
This commit adds the `isDemoUser` middleware to the `auth.middleware.js` file. The middleware checks if the user's role is set to 'demo' and throws a `BadRequestError` if it is. This enhancement improves the security of the application by preventing demo users from performing actions that are not supported in the demo mode.

Note: This commit message follows the established conventions in the repository.