feat: Add AreaChart component for displaying data in an area chart
This commit adds the AreaChart component to the frontend/src/components/AreaChart.jsx file. The component is responsible for rendering data in the form of an area chart using the recharts library. It accepts a data prop, which should be an array of objects containing the date and count properties. The AreaChart component enhances the application by providing a visually appealing and informative representation of the data.