build: Update main.jsx to add react-query and react-query-devtools
This commit updates the main.jsx file in the frontend/src directory to add the react-query and react-query-devtools dependencies. The QueryClientProvider and ReactQueryDevtools components are also added to wrap the application, enabling the usage of react-query for data fetching and caching. This change improves the application's data management capabilities.

Note: This commit message follows the established conventions in the repository.