feat: Add ErrorElement component to Stats route
The main.jsx file in the frontend/src directory has been updated to import and render the ErrorElement component in the Stats route. This change ensures that an error element is displayed when there is an error loading the stats data. The addition of the ErrorElement component enhances the user experience by providing visual feedback in case of errors.

Note: This commit message follows the established conventions in the repository.